+++ internal/install/types/types.go
@@ -205,7 +205,8 @@ type OpenInstallationRecipeInstallTarget struct {
 // See the [NRQL Docs](https://docs.newrelic.com/docs/insights/nrql-new-relic-query-language/nrql-resources/nrql-syntax-components-functions) for more information about NRQL syntax.
 type NRQL string
 
-type SuccessLink struct {
+// SuccessLinkConfig contains metadata to support generating a URL after installation success
+type SuccessLinkConfig struct {
 	Type   string `json:"type"`
 	Filter string `json:"filter"`
 }