+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+<a name="v0.20.7"></a>
+## [v0.20.7] - 2021-03-31
+### Bug Fixes
+- **install:** update entity URL with region
+
+<a name="v0.20.6"></a>
+## [v0.20.6] - 2021-03-31
+### Bug Fixes
+- **install:** fetch license key when installing
+- **install:** fetch license key when installing
+- **install:** fetch license key when installing
+
+<a name="v0.20.5"></a>
+## [v0.20.5] - 2021-03-29
+### Features
+- **install:** additional queryable fields for InstallStatus event
+
+<a name="v0.20.4"></a>
+## [v0.20.4] - 2021-03-24
+### Bug Fixes
+- avoid nil pointer when fetching a license key
+
+<a name="v0.20.3"></a>
+## [v0.20.3] - 2021-03-23
+### Features
+- **install:** enable stdin piping for install command
+
 <a name="v0.20.2"></a>
 ## [v0.20.2] - 2021-03-18
 ### Bug Fixes
@@ -408,8 +435,6 @@
 - **output:** Output package for central output handling
 - **snapcraft:** include goreleaser config for snaps
 
-<a name="v0.6.2-test"></a>
-## [v0.6.2-test] - 2020-04-09
 <a name="v0.6.2"></a>
 ## [v0.6.2] - 2020-04-08
 ### Bug Fixes
@@ -543,7 +568,12 @@
 - **profile:** Enable reading of profiles and use Region/APIKey from default profile
 - **profile:** Add listing of profiles to command
 
-[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.2...HEAD
+[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.7...HEAD
+[v0.20.7]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.6...v0.20.7
+[v0.20.6]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.5...v0.20.6
+[v0.20.5]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.4...v0.20.5
+[v0.20.4]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.3...v0.20.4
+[v0.20.3]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.2...v0.20.3
 [v0.20.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.1...v0.20.2
 [v0.20.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.20.0...v0.20.1
 [v0.20.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.19.2...v0.20.0
@@ -604,8 +634,7 @@
 [v0.8.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.8.1...v0.8.2
 [v0.8.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.8.0...v0.8.1
 [v0.8.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.7.0...v0.8.0
-[v0.7.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.6.2-test...v0.7.0
-[v0.6.2-test]: https://github.com/newrelic/newrelic-client-go/compare/v0.6.2...v0.6.2-test
+[v0.7.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.6.2...v0.7.0
 [v0.6.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.6.1...v0.6.2
 [v0.6.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.6.0...v0.6.1
 [v0.6.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.5.0...v0.6.0